Rename stuff, add tax to table
