+++ src/routes/shipments/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import Link from '$lib/components/Link.svelte';
 	import TableWithHeader from '$lib/components/TableWithHeader.svelte';
-	import StatusBadge from '$lib/components/status-badge.svelte';
+	import StatusBadge from '$lib/components/StatusBadge.svelte';
 
 	let { data } = $props();
 	let shipments = $derived(data.shipments);
